Compile with Java 11, describe release property
Only setting source and target is not sufficient to produce bytecode
that is properly compatible with Java 8. The new release flag also sets
the bootclasspath so that everything is set up properly.

However, this will need Java 11 for building. So for now I'm not
enabling the option altogether but only documenting it.

In addition, bump to the latest compile plugin and surefire to ensure
Java 11 compatibility.