fix NoSuchMethodError with checkstyle
dependabot upgraded the `com.puppycrawl.tools:checkstyle` dependency of
the maven-checkstyle-plugin in eac05c272 which results in
incompatabilities between the plugin and Checkstyle itself:

> [ERROR] Failed to execute goal org.apache.maven.plugins:maven-checkstyle-plugin:3.0.0:check (validate) on project logging: Execution validate of goal org.apache.maven.plugins:maven-checkstyle-plugin:3.0.0:check failed: An API incompatibility was encountered while executing org.apache.maven.plugins:maven-checkstyle-plugin:3.0.0:check: java.lang.NoSuchMethodError: 'void com.puppycrawl.tools.checkstyle.DefaultLogger.<init>(java.io.OutputStream, boolean)'

This changes bumps maven-checkstyle-plugin to the latest version (3.1.1)
to fix this. Since the plugin now uses checkstyle 8.29, there is not a
need to pin the version in the plugin's `<dependencies>`. Its also
necessary to update the version of `spotify-checkstyle-config` as the
format of the checkstyle config file has changed.